refactor: separate Built-in target list from FsmTextHook
